+++ Changelog.md
@@ -1,5 +1,9 @@
 OpenCore Changelog
 ==================
+#### v0.7.5
+- Revised OpenLinuxBoot documentation
+- Supported Linux ostree boot layout
+
 #### v0.7.4
 - Fixed Linux kernel sort order
 - Added Linux detection optional log detail